Add dynamic path logic for editor and player, nice :)

+++ Game/src/game_objects/board_space/BoardSpace.h
@@ -7,9 +7,9 @@ class BoardSpace : public GameObject {
 		BoardSpace(uint16_t id, Model* model, glm::vec3 pos, glm::vec3 size, glm::vec3 rotation);
 		BoardSpace(GameObject* gameObject);
 		//virtual void process_space() = 0;
-
+		
 		std::vector<BoardSpace*> nextSpace;
-		std::vector<uint8_t> nextSpaceIds;
+		std::vector<int> nextSpaceIds;
 		uint8_t spaceType = 0;
 };
